refactor: move conversion code to utils
